requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.0
openai==1.57.4
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.9
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==74.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
